fix(backend): fix code after updating Python package dependencies
- `cherrypy` was split into two projects. Now we are using `cheroot`
- added dependency from `types-requests` for static analysis

+++ lib_common.py
@@ -6,5 +6,5 @@ import os
 import stat
 
 
-def get_file_modified_unixtime(pathname: str) -> int:
+def get_file_modified_unixtime(pathname: str) -> float:
     return os.stat(pathname)[stat.ST_MTIME]
